+++ lib/ansible/inventory/expand_hosts.py
@@ -1,6 +1,5 @@
-#
-# $Id: expand_hosts.py,v 1.9 2012/07/22 16:17:57 fangchin Exp $
-#
+#!/usr/bin/python
+
 # (c) 2012, Zettar Inc.
 # Written by Chin Fang <fangchin@zettar.com>
 #
@@ -22,8 +21,15 @@
 '''
 This module is for enhancing ansible's inventory parsing capability such
 that it can deal with hostnames specified using a simple pattern in the
-form of [beg:end:step], example: [1:5:2] where if beg is not specified, it
-defaults to 0. If step is not specified, it defaults to 1.
+form of [beg:end], example: [1:5] where if beg is not specified, it
+defaults to 0.
+
+If beg is given and is left-zero-padded, e.g. '001', it is taken as a
+formatting hint when the range is expanded. e.g. [001:010] is to be
+expanded into 001, 002 ...009, 010.
+
+Note that when beg is specified with left zero padding, then the length of
+end must be the same as that of beg, else a exception is raised.
 '''
 import sys
 from pprint import pprint
@@ -31,14 +37,7 @@ from pprint import pprint
 def detect_range(line = None):
     '''
     A helper function that checks a given host line to see if it contains
-    a range pattern. The following are examples:
-
-    o node[1:6]
-    o node[1:6:2]
-    o node[1:6].example.com
-    o node[1:6:2].example.com
-    o node[1:6]-webserver
-    o node[1:6:2]-database
+    a range pattern descibed in the docstring above.
 
     Returnes True if the given line contains a pattern, else False.
     '''
@@ -54,8 +53,8 @@ def detect_range(line = None):
 def expand_hostname_range(line = None):
     '''
     A helper function that expands a given line that contains a pattern
-    specified in def detect_range, and returns a list that consists
-    of the expanded version.
+    specified in top docstring, and returns a list that consists of the
+    expanded version.
 
     The '[' and ']' characters are used to maintain the pseudo-code
     appearance. They are replaced in this function with '|' to ease
@@ -65,36 +64,32 @@ def expand_hostname_range(line = None):
     '''
     all_hosts = []
     if line:
-        # A hostname such as db[1:6:2]-node is considered to consists
+        # A hostname such as db[1:6]-node is considered to consists
         # three parts: 
         # head: 'db'
-        # nrange: [1:6:2]; range is a built-in. Can't use the name
+        # nrange: [1:6]; range() is a built-in. Can't use the name
         # tail: '-node'
         
         (head, nrange, tail) = line.replace('[','|').replace(']','|').split('|')
         bounds = nrange.split(":")
-        lbounds = len(bounds)
+        if len(bounds) != 2:
+            raise ValueError("host range incorrectly specified!")
         beg = bounds[0]
         end = bounds[1]
-        lbounds = len(bounds)
-        if lbounds == 2:
-            step = 1
-        elif lbounds == 3:
-            step = bounds[2]
-        else:
-            raise ValueError("host range incorrectly specified!")
-        
         if not beg:
             beg = "0"
-            
         if not end:
-            raise ValueError("host range incorrectly specified!")
-            
-        if not step:
-            step = "1"
-               
-        for _ in range(int(beg), int(end), int(step)):
-            hname = ''.join((head, str(_), tail))
+            raise ValueError("host range end value missing!")
+        rlen = len(beg) # range length formatting hint
+        if rlen > 1 and rlen != len(end):
+            raise ValueError("host range format incorrectly specified!")
+                
+        for _ in range(int(beg), int(end)):
+            if rlen > 1:
+                rseq = str(_).zfill(rlen) # range sequence
+            else:
+                rseq = str(_)
+            hname = ''.join((head, rseq, tail))
             all_hosts.append(hname)
                    
         return all_hosts
@@ -113,13 +108,11 @@ def main():
                  "db.example.com",
                  "node[:6]", 
                  "node[1:6]",
-                 "node[1:6:2]",
+                 "node[001:006]",
                  "node[:5].example.com",
-                 "node[1:6].example.com",
-                 "node[1:6:2].example.com",
+                 "node-[01:06].example.com",
                  "node[:6]-webserver", 
-                 "node[1:6]-webserver",
-                 "node[1:6:2]-webserver"]
+                 "node-[0001:0006]-webserver"]
 
     for data in test_data:
         print "===> testing %s..." % data